(lib) move state includes internally
The externally-visible API needs not know about the internal structure
of the Casemate model state.

This moves the state typedefs out of casemate.h and into casemate-impl.h
and makes the initialisation function error out if not supplied enough space.

At the same time, drop to the standard (u)intN_s in the API header.

+++ examples/common.c
@@ -313,7 +313,7 @@ void common_init(int argc, char **argv)
 	opts.enable_tracing = SHOULD_TRACE;
 
 	st = malloc(sm_size);
-	initialise_casemate_model(&opts, 0, 0, (u64)st, sm_size);
+	initialise_casemate_model(&opts, 0, 0, st, sm_size);
 	initialise_ghost_driver(&sm_driver);
 
 	mtx_init(&m, mtx_plain);
